requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.0
asttokens==3.0.0
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==45.0.5
decorator==5.2.1
Django==5.2.1
django-celery-beat==2.8.1
django-phonenumber-field==8.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.10
executing==2.2.0
firebase-admin==6.9.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
ipython==4.0.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
msgpack==1.1.1
numpy==2.2.3
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
phonenumbers==9.0.9
pickleshare==0.7.5
pillow==11.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rsa==4.9.1
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
